requirements.txt
google-generativeai
openai
ruff
isort
